@@ -1,19 +1,18 @@
----
 name: flutter_iconpicker
 description: A Dialog for picking Icons in Flutter and use them anywhere. Can be
   used as a default Dialog or as a Adaptive Dialog.
-version: 3.2.1
+version: 3.2.2
 homepage: https://github.com/Ahmadre
 repository: https://github.com/Ahmadre/FlutterIconPicker
 environment:
-  sdk: ">=2.13.0 <3.0.0"
-  flutter: ">=2.2.0"
+  sdk: '>=2.13.0 <3.0.0'
+  flutter: '>=3.0.0'
 dependencies:
-  cupertino_icons: ^1.0.4
+  cupertino_icons: ^1.0.5
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.1.0
-  provider: ^6.0.3
+  font_awesome_flutter: ^10.2.1
+  provider: ^6.0.4
 dev_dependencies:
   flutter_test:
     sdk: flutter
