@@ -8,20 +8,20 @@ import 'app_brightness.dart';
 
 class IconNotifier extends ChangeNotifier {
   IconNotifier._(
-    IconData iconData,
+    IconData? iconData,
     AppBrightness brightness,
   )   : _iconData = iconData,
         _brightness = brightness;
 
-  static Box box;
+  static late Box box;
 
-  IconData _iconData;
+  IconData? _iconData;
 
   AppBrightness _brightness;
 
-  IconData get iconData => _iconData;
+  IconData? get iconData => _iconData;
 
-  set iconData(IconData value) {
+  set iconData(IconData? value) {
     if (_iconData == value) {
       return;
     }
@@ -34,7 +34,7 @@ class IconNotifier extends ChangeNotifier {
       return;
     }
 
-    box.put('iconData', serializeIcon(_iconData));
+    box.put('iconData', serializeIcon(_iconData!));
     notifyListeners();
   }
 
